+++ plugins/postcss-focus-within/src/index.ts
@@ -2,7 +2,7 @@ import parser from 'postcss-selector-parser';
 import type { PluginCreator } from 'postcss';
 import isValidReplacement from './is-valid-replacement.mjs';
 
-type pluginOptions = { preserve?: boolean, replaceWith?: string };
+type pluginOptions = { preserve?: boolean, replaceWith?: string, disablePolyfillReadyClass?: boolean };
 
 const POLYFILL_READY_CLASSNAME = 'js-focus-within';
 const PSEUDO = ':focus-within';
@@ -13,6 +13,7 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 		{
 			preserve: true,
 			replaceWith: '[focus-within]',
+			disablePolyfillReadyClass: false,
 		},
 		// Provided options
 		opts,
@@ -80,7 +81,7 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 				// becomes:
 				// html.js-focus-within > .foo:focus-within,
 				// .js-focus-within html > .foo:focus-within
-				{
+				if (!options.disablePolyfillReadyClass) {
 					if (selectorAST.nodes?.[0]?.nodes?.length) {
 						for (let i = 0; i < selectorAST.nodes[0].nodes.length; i++) {
 							const node = selectorAST.nodes[0].nodes[i];